+++ aleph/model/alert.py
@@ -3,7 +3,7 @@ from werkzeug.datastructures import MultiDict
 from sqlalchemy import func
 
 from aleph.core import db
-from aleph.data.validation import validate
+from aleph.data.validate import validate
 from aleph.model.entity import Entity
 from aleph.model.common import SoftDeleteModel
