+++ aleph/model/collection.py
@@ -5,7 +5,7 @@ from sqlalchemy.orm import aliased
 from sqlalchemy.dialects.postgresql import ARRAY
 
 from aleph.core import db, url_for
-from aleph.data.validation import validate
+from aleph.data.validate import validate
 from aleph.model.role import Role
 from aleph.model.permission import Permission
 from aleph.model.util import ModelFacets
