+++ aleph/model/entity.py
@@ -6,7 +6,7 @@ from sqlalchemy.dialects.postgresql import JSONB
 
 from aleph.core import db
 from aleph.text import normalize_strong
-from aleph.data.validation import validate
+from aleph.data.validate import validate
 from aleph.model.collection import Collection
 from aleph.model.reference import Reference
 from aleph.model.common import SoftDeleteModel, UuidModel
