+++ aleph/views/base_api.py
@@ -9,7 +9,7 @@ from elasticsearch import TransportError
 from aleph.core import get_config, app_title, app_url
 from aleph.metadata import Metadata
 from aleph.reference import COUNTRY_NAMES, LANGUAGE_NAMES
-from aleph.data.validation import resolver
+from aleph.data.validate import resolver
 from aleph.views.cache import enable_cache
 
 blueprint = Blueprint('base_api', __name__)
